grant

+++ client.py
@@ -79,20 +79,20 @@ def checkin():
 	'''
 	did=raw_input("Enter unique document name (1.txt):")
 	flag=raw_input("Enter security flag (1-confidentiality, 2-integrity):")
-	try:
+	if os.path.exists("documents/checkin/"+did): 
+		print "ioerror"
 		fin = open('documents/checkin/'+did, 'r')
-	except IOError:
-		try:
-			print "ioerror 1"
-			fin = open('documents/checkout/'+did, 'r')
-			f = open("documents/checkoin/"+did,"w")
-			f.write(fin.read())
-			f.close()
-			print "ioerror 1"
-			os.remove("documents/checkout/"+did) 
-		except IOError:
-			print "Document does not exist on client side."
-			return
+	else if os.path.exists("documents/checkout/"+did):
+		print "ioerror 1"
+		fin = open('documents/checkout/'+did, 'r')
+		f = open("documents/checkin/"+did,"w")
+		f.write(fin.read())
+		f.close()
+		print "ioerror 1"
+		os.remove("documents/checkout/"+did) 
+	else:
+		print "Document does not exist on client side."
+		return
 
 	data={
 		'did':did,
